Added tabgui

+++ src/main/java/net/minecraft/client/gui/Gui.java
@@ -68,15 +68,15 @@ public class Gui {
      * Draws a solid color rectangle with the specified coordinates
      * and color (ARGB format). Args: x1, y1, x2, y2, color
      */
-    public static void drawRect(int left, int top, int right, int bottom, int color) {
+    public static void drawRect(double left, double top, double right, double bottom, int color) {
         if (left < right) {
-            int i = left;
+            double i = left;
             left = right;
             right = i;
         }
 
         if (top < bottom) {
-            int j = top;
+            double j = top;
             top = bottom;
             bottom = j;
         }
